Move get_or_create_user to test code
This is only used in tests, so...
